requirements.txt
docopt
pyyaml
oauth2client
google-auth
google-api-python-client
